requirements-dev.txt
-r requirements.txt

ruff
coverage
django-coverage
model-bakery